requirements.txt
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
websockets>=12.0

# ML dependencies for heating optimization
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
